+++ src/routes/v2/projects/[projectId]/workflows/[workflowId]/+page.svelte
@@ -6,7 +6,11 @@
 	import ConfirmActionButton from '$lib/components/common/ConfirmActionButton.svelte';
 	import MetaPropertiesForm from '$lib/components/v2/workflow/MetaPropertiesForm.svelte';
 	import ArgumentsSchema from '$lib/components/v2/workflow/ArgumentsSchema.svelte';
-	import { displayStandardErrorAlert, getAlertErrorFromResponse } from '$lib/common/errors';
+	import {
+		displayStandardErrorAlert,
+		FormErrorHandler,
+		getAlertErrorFromResponse
+	} from '$lib/common/errors';
 	import Modal from '$lib/components/common/Modal.svelte';
 	import StandardDismissableAlert from '$lib/components/common/StandardDismissableAlert.svelte';
 	import VersionUpdate from '$lib/components/v2/workflow/VersionUpdate.svelte';
@@ -14,7 +18,7 @@
 	import TasksOrderModal from '$lib/components/v2/workflow/TasksOrderModal.svelte';
 	import { extractRelevantJobError, showExecutorErrorLog } from '$lib/common/job_utilities';
 	import JobLogsModal from '$lib/components/v2/jobs/JobLogsModal.svelte';
-	import TaskInfoTab from '$lib/components/v2/workflow/TaskInfoTab.svelte';
+	import WorkflowTaskInfoTab from '$lib/components/v2/workflow/WorkflowTaskInfoTab.svelte';
 	import InputFiltersTab from '$lib/components/v2/workflow/InputFiltersTab.svelte';
 	import RunWorkflowModal from '$lib/components/v2/workflow/RunWorkflowModal.svelte';
 	import { getSelectedWorkflowDataset, saveSelectedDataset } from '$lib/common/workflow_utilities';
@@ -88,6 +92,7 @@
 
 	// Update workflow modal
 	let updatedWorkflowName = $state('');
+	let updatedWorkflowDescription = $state('');
 
 	// Modals
 	/** @type {Modal|undefined} */
@@ -113,6 +118,14 @@
 	/** @type {import('fractal-components/types/api').ApplyWorkflowV2|undefined} */
 	let selectedSubmittedJob = $state();
 	let jobCancelledMessage = $state('');
+
+	let showWorkflowDescription = $state(false);
+
+	const workflowPropsErrorHandler = new FormErrorHandler('errorAlert-editWorkflowModal', [
+		'name',
+		'description'
+	]);
+	const workflowPropsValidationErrors = workflowPropsErrorHandler.getValidationErrorStore();
 
 	let updatableWorkflowList = $derived(workflow.task_list || []);
 
@@ -220,6 +233,7 @@
 			return;
 		}
 		updatedWorkflowName = workflow.name;
+		updatedWorkflowDescription = workflow.description || '';
 	}
 
 	let workflowUpdating = $state(false);
@@ -229,38 +243,40 @@
 	 * @returns {Promise<void>}
 	 */
 	async function handleWorkflowUpdate() {
-		editWorkflowModal?.confirmAndHide(
-			async () => {
-				workflowSuccessMessage = '';
-				if (!workflow) {
-					return;
-				}
-				workflowUpdating = true;
+		workflowSuccessMessage = '';
+		workflowPropsErrorHandler.clearErrors();
 
-				const headers = new Headers();
-				headers.set('Content-Type', 'application/json');
+		if (!workflow) {
+			return;
+		}
+		workflowUpdating = true;
 
-				const response = await fetch(`/api/v2/project/${project.id}/workflow/${workflow.id}`, {
-					method: 'PATCH',
-					credentials: 'include',
-					headers,
-					body: normalizePayload({
-						name: updatedWorkflowName
-					})
-				});
+		const headers = new Headers();
+		headers.set('Content-Type', 'application/json');
 
-				if (response.ok) {
-					workflow = await response.json();
-					workflowSuccessMessage = 'Workflow updated correctly';
-				} else {
-					console.error('Error updating workflow properties');
-					throw await getAlertErrorFromResponse(response);
-				}
-			},
-			() => {
-				workflowUpdating = false;
-			}
-		);
+		const response = await fetch(`/api/v2/project/${project.id}/workflow/${workflow.id}`, {
+			method: 'PATCH',
+			credentials: 'include',
+			headers,
+			body: normalizePayload(
+				{
+					name: updatedWorkflowName,
+					description: updatedWorkflowDescription
+				},
+				{ nullifyEmptyStrings: true }
+			)
+		});
+
+		if (response.ok) {
+			workflow = await response.json();
+			workflowSuccessMessage = 'Workflow updated correctly';
+			editWorkflowModal?.hide();
+		} else {
+			console.error('Error updating workflow properties');
+			await workflowPropsErrorHandler.handleErrorResponse(response);
+		}
+
+		workflowUpdating = false;
 	}
 
 	/**
@@ -686,8 +702,8 @@
 	});
 </script>
 
-<div class="container mt-3">
-	<nav aria-label="breadcrumb">
+<div class="container mt-3 d-flex">
+	<nav aria-label="breadcrumb" class="flex-grow-1">
 		<ol class="breadcrumb">
 			<li class="breadcrumb-item" aria-current="page">
 				<a href="/v2/projects">Projects</a>
@@ -705,9 +721,31 @@
 			{/if}
 		</ol>
 	</nav>
+	<div>
+		{#if workflow.description}
+			<div class="form-check form-switch float-end">
+				<input
+					class="form-check-input"
+					bind:checked={showWorkflowDescription}
+					type="checkbox"
+					role="switch"
+					id="showWorkflowDescriptionSwitch"
+				/>
+				<label class="form-check-label" for="showWorkflowDescriptionSwitch">
+					Show description
+				</label>
+			</div>
+		{/if}
+	</div>
 </div>
 
-<div class="container mt-2">
+<div class="container">
+	{#if workflow.description && showWorkflowDescription}
+		<div class="mb-3">
+			{workflow.description}
+		</div>
+	{/if}
+
 	<div class="row">
 		<div class="col-lg-8">
 			<div class="row">
@@ -812,7 +850,7 @@
 <TypeFiltersFlowModal {workflow} bind:this={typeFiltersFlowModal} />
 
 {#if workflow}
-	<div class="container mt-2">
+	<div class="container">
 		<StandardDismissableAlert message={jobCancelledMessage} />
 		<StandardDismissableAlert message={workflowSuccessMessage} />
 
@@ -912,7 +950,7 @@
 											</button>
 										{/if}
 									{/if}
-									{workflowTask.task.name}
+									{workflowTask.alias ? workflowTask.alias : workflowTask.task.name}
 									<span class="float-end ps-2">
 										{#if selectedDataset}
 											{#if isLegacy}
@@ -1127,7 +1165,16 @@
 							<div id="info-tab" class="tab-pane show active">
 								<div class="card-body">
 									{#if selectedWorkflowTask}
-										<TaskInfoTab task={selectedWorkflowTask.task} />
+										<WorkflowTaskInfoTab
+											projectId={workflow.project_id}
+											workflowTask={selectedWorkflowTask}
+											updateWorkflowTaskCallback={(up) => {
+												selectedWorkflowTask = up;
+												workflow.task_list = workflow.task_list.map((wt) =>
+													wt.id === up.id ? up : wt
+												);
+											}}
+										/>
 									{/if}
 								</div>
 							</div>
@@ -1176,7 +1223,6 @@
 		<h5 class="modal-title">Workflow properties</h5>
 	{/snippet}
 	{#snippet body()}
-		<div id="errorAlert-editWorkflowModal"></div>
 		{#if workflow}
 			<form
 				id="updateWorkflow"
@@ -1184,18 +1230,33 @@
 					e.preventDefault();
 					handleWorkflowUpdate();
 				}}
+				class="has-validation"
 			>
 				<div class="mb-3">
 					<label for="workflowName" class="form-label">Workflow name</label>
 					<input
 						type="text"
 						class="form-control"
+						class:is-invalid={$workflowPropsValidationErrors['name']}
 						name="workflowName"
 						id="workflowName"
 						bind:value={updatedWorkflowName}
 					/>
+					<span class="invalid-feedback">{$workflowPropsValidationErrors['name']}</span>
+				</div>
+				<div class="mb-3">
+					<label for="workflowDescription" class="form-label">Workflow description</label>
+					<textarea
+						class="form-control"
+						class:is-invalid={$workflowPropsValidationErrors['description']}
+						name="workflowDescription"
+						id="workflowDescription"
+						bind:value={updatedWorkflowDescription}
+					></textarea>
+					<span class="invalid-feedback">{$workflowPropsValidationErrors['description']}</span>
 				</div>
 			</form>
+			<div id="errorAlert-editWorkflowModal"></div>
 		{/if}
 	{/snippet}
 	{#snippet footer()}