+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added the option to customize the login invite, for example to add a disclaimer (\#934).
 * Fixed typo in UI (\#934).
 * Fixed issue to get correct URL for s3 zarr paths (\#935).
+* Supported richer workflow/workflowtask metadata (\#938).
 
 # 1.22.2